# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Firesky Industries Field Operations API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DashboardSummaryJobsByStage = { [key: string]: number };

export type EnquiryStatus = (typeof EnquiryStatus)[keyof typeof EnquiryStatus];

export const EnquiryStatus = {
  new: "new",
  in_progress: "in_progress",
  inspection_done: "inspection_done",
  quoted: "quoted",
  won: "won",
  lost: "lost",
} as const;

export type EnquiryPriority =
  (typeof EnquiryPriority)[keyof typeof EnquiryPriority];

export const EnquiryPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Enquiry {
  id: number;
  customerId: number;
  customerName?: string;
  title: string;
  description?: string;
  tankSize?: string;
  tankQuantity?: number;
  status: EnquiryStatus;
  priority?: EnquiryPriority;
  notes?: string;
  createdAt: string;
  updatedAt: string;
}

export type JobStage = (typeof JobStage)[keyof typeof JobStage];

export const JobStage = {
  enquiry: "enquiry",
  inspection: "inspection",
  quoting: "quoting",
  quoted: "quoted",
  won: "won",
  lost: "lost",
} as const;

export type JobPriority = (typeof JobPriority)[keyof typeof JobPriority];

export const JobPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Job {
  id: number;
  customerId: number;
  customerName?: string;
  enquiryId?: number;
  inspectionId?: number;
  title: string;
  stage: JobStage;
  priority?: JobPriority;
  tankSize?: string;
  tankQuantity?: number;
  estimatedValue?: number;
  notes?: string;
  createdAt: string;
  updatedAt: string;
}

export interface DashboardSummary {
  totalCustomers: number;
  totalEnquiries: number;
  totalJobs: number;
  jobsByStage: DashboardSummaryJobsByStage;
  recentEnquiries: Enquiry[];
  recentJobs: Job[];
}

export interface Customer {
  id: number;
  name: string;
  contactName?: string;
  phone?: string;
  email?: string;
  farmName?: string;
  nearestTown?: string;
  province?: string;
  manualDirections?: string;
  landmarks?: string;
  whatsappLocation?: string;
  accessNotes?: string;
  notes?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCustomerBody {
  name: string;
  contactName?: string;
  phone?: string;
  email?: string;
  farmName?: string;
  nearestTown?: string;
  province?: string;
  manualDirections?: string;
  landmarks?: string;
  whatsappLocation?: string;
  accessNotes?: string;
  notes?: string;
}

export type CreateEnquiryBodyStatus =
  (typeof CreateEnquiryBodyStatus)[keyof typeof CreateEnquiryBodyStatus];

export const CreateEnquiryBodyStatus = {
  new: "new",
  in_progress: "in_progress",
  inspection_done: "inspection_done",
  quoted: "quoted",
  won: "won",
  lost: "lost",
} as const;

export type CreateEnquiryBodyPriority =
  (typeof CreateEnquiryBodyPriority)[keyof typeof CreateEnquiryBodyPriority];

export const CreateEnquiryBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateEnquiryBody {
  customerId: number;
  title: string;
  description?: string;
  tankSize?: string;
  tankQuantity?: number;
  status?: CreateEnquiryBodyStatus;
  priority?: CreateEnquiryBodyPriority;
  notes?: string;
}

export interface Inspection {
  id: number;
  enquiryId?: number;
  customerId: number;
  customerName?: string;
  farmName?: string;
  nearestTown?: string;
  manualDirections?: string;
  landmarks?: string;
  whatsappLocation?: string;
  accessNotes?: string;
  tankSize?: string;
  tankQuantity?: number;
  requiresStand?: boolean;
  requiresPlinth?: boolean;
  standHeight?: string;
  plinthDetails?: string;
  pipeLength?: number;
  pipeDetails?: string;
  distanceFromRoad?: number;
  distanceFromHouse?: number;
  truckAccess?: boolean;
  trailerAccess?: boolean;
  offloadingConstraints?: string;
  groundCondition?: string;
  siteReadyToQuote?: boolean;
  photoUrls?: string[];
  notes?: string;
  inspectedAt?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateInspectionBody {
  enquiryId?: number;
  customerId: number;
  farmName?: string;
  nearestTown?: string;
  manualDirections?: string;
  landmarks?: string;
  whatsappLocation?: string;
  accessNotes?: string;
  tankSize?: string;
  tankQuantity?: number;
  requiresStand?: boolean;
  requiresPlinth?: boolean;
  standHeight?: string;
  plinthDetails?: string;
  pipeLength?: number;
  pipeDetails?: string;
  distanceFromRoad?: number;
  distanceFromHouse?: number;
  truckAccess?: boolean;
  trailerAccess?: boolean;
  offloadingConstraints?: string;
  groundCondition?: string;
  siteReadyToQuote?: boolean;
  photoUrls?: string[];
  notes?: string;
  inspectedAt?: string;
}

export type CreateJobBodyStage =
  (typeof CreateJobBodyStage)[keyof typeof CreateJobBodyStage];

export const CreateJobBodyStage = {
  enquiry: "enquiry",
  inspection: "inspection",
  quoting: "quoting",
  quoted: "quoted",
  won: "won",
  lost: "lost",
} as const;

export type CreateJobBodyPriority =
  (typeof CreateJobBodyPriority)[keyof typeof CreateJobBodyPriority];

export const CreateJobBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateJobBody {
  customerId: number;
  enquiryId?: number;
  inspectionId?: number;
  title: string;
  stage: CreateJobBodyStage;
  priority?: CreateJobBodyPriority;
  tankSize?: string;
  tankQuantity?: number;
  estimatedValue?: number;
  notes?: string;
}

export type ListCustomersParams = {
  search?: string;
};

export type ListEnquiriesParams = {
  customerId?: number;
  status?: string;
};

export type ListInspectionsParams = {
  enquiryId?: number;
  customerId?: number;
};

export type ListJobsParams = {
  stage?: string;
  customerId?: number;
};