# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Firesky Industries Field Operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /enquiries:
    get:
      operationId: listEnquiries
      tags: [enquiries]
      summary: List all enquiries
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of enquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enquiry"
    post:
      operationId: createEnquiry
      tags: [enquiries]
      summary: Create an enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnquiryBody"
      responses:
        "201":
          description: Created enquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
  /enquiries/{id}:
    get:
      operationId: getEnquiry
      tags: [enquiries]
      summary: Get an enquiry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
    put:
      operationId: updateEnquiry
      tags: [enquiries]
      summary: Update an enquiry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnquiryBody"
      responses:
        "200":
          description: Updated enquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactName:
          type: string
        phone:
          type: string
        email:
          type: string
        farmName:
          type: string
        nearestTown:
          type: string
        province:
          type: string
        manualDirections:
          type: string
        landmarks:
          type: string
        whatsappLocation:
          type: string
        accessNotes:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
        - updatedAt
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        contactName:
          type: string
        phone:
          type: string
        email:
          type: string
        farmName:
          type: string
        nearestTown:
          type: string
        province:
          type: string
        manualDirections:
          type: string
        landmarks:
          type: string
        whatsappLocation:
          type: string
        accessNotes:
          type: string
        notes:
          type: string
      required:
        - name
    Enquiry:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        title:
          type: string
        description:
          type: string
        tankSize:
          type: string
        tankQuantity:
          type: integer
        status:
          type: string
          enum: [new, in_progress, inspection_done, quoted, won, lost]
        priority:
          type: string
          enum: [low, medium, high]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - title
        - status
        - createdAt
        - updatedAt
    CreateEnquiryBody:
      type: object
      properties:
        customerId:
          type: integer
        title:
          type: string
        description:
          type: string
        tankSize:
          type: string
        tankQuantity:
          type: integer
        status:
          type: string
          enum: [new, in_progress, inspection_done, quoted, won, lost]
        priority:
          type: string
          enum: [low, medium, high]
        notes:
          type: string
      required:
        - customerId
        - title
...