# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Firesky Industries Field Operations API
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  id: number;
  name: string;
  contactName?: string;
  phone?: string;
  email?: string;
  farmName?: string;
  nearestTown?: string;
  province?: string;
  manualDirections?: string;
  landmarks?: string;
  whatsappLocation?: string;
  accessNotes?: string;
  notes?: string;
  createdAt: Date;
  updatedAt: Date;
}