# lib/api-zod/src/generated/types/enquiry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Firesky Industries Field Operations API
 * OpenAPI spec version: 0.1.0
 */
import type { EnquiryPriority } from "./enquiryPriority";
import type { EnquiryStatus } from "./enquiryStatus";

export interface Enquiry {
  id: number;
  customerId: number;
  customerName?: string;
  title: string;
  description?: string;
  tankSize?: string;
  tankQuantity?: number;
  status: EnquiryStatus;
  priority?: EnquiryPriority;
  notes?: string;
  createdAt: Date;
  updatedAt: Date;
}